+++ connection/src/main/java/io/herow/sdk/connection/cache/model/mapper/PoiMapper.kt
@@ -2,6 +2,6 @@ package io.herow.sdk.connection.cache.model.mapper
 
 data class PoiMapper(
     var id: String = "",
-    var distance: Float = 0f,
+    var distance: Double = 0.0,
     var tags: List<String>? = listOf()
 )
